Remove: 불필요한 파일 삭제 - amountHandler 컴포넌트 파일 삭제
